chore(friends): rename folder name friendlist to friends and remove conflict method
